+++ backend/audiosm/tasks/stream.py
@@ -1,5 +1,5 @@
 from audiosm.extensions import celery
-from ffmpy import FFmpeg
+from ffmpy import FFmpeg, FFRuntimeError
 from audiosm import settings
 from audiosm.extensions import DatabaseTask
 from audiosm.jobs.models import JobModel
@@ -15,24 +15,37 @@ import subprocess
 
 
 @celery.task(ignore_result=True, base=DatabaseTask, bind=True)
-def play_audio_task(self, id, filename, stream, inf, *a, **kw):
+def play_audio_task(self, id, files, stream, inf, *a, **kw):
     loop = '-1' if inf else '0'
-    f = FFmpeg(
-        inputs={
-            filename: ['-re', '-stream_loop', loop]
-        },
-        outputs={
-            'rtmp://{host}:1935/stream/{stream}'.format(
-                stream=stream['name'],
-                host=settings.Config.STREAM_HOST
-            ): ['-vn', '-c:a', 'aac', '-strict',  '-2', '-f', 'flv']
-        }
-    )
-    print('start')
-    print("FFmpeg command", f.cmd)
-    print("Running Command")
-    f.run(stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    problem = None
+    for filename in files:
+        f = FFmpeg(
+            inputs={
+                filename: ['-re', '-stream_loop', loop]
+            },
+            outputs={
+                'rtmp://{host}:1935/stream/{stream}'.format(
+                    stream=stream['name'],
+                    host=settings.Config.STREAM_HOST
+                ): ['-vn', '-c:a', 'aac', '-strict',  '-2', '-f', 'flv']
+            }
+        )
+        print('start')
+        print("FFmpeg command", f.cmd)
+        print("Running Command")
+        try:
+            f.run(stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        except FFRuntimeError as e:
+            problem = e
+            continue
+
+    if problem:
+        print("An error has occured when running job")
+        print(problem)
     job = self.db.session.query(JobModel).filter_by(id=id).first()
     print("Finish running Command")
-    print("Deleting job", job)
-    job.delete()
+    if job:
+        print("Deleting job", job)
+        job.delete()
+    else:
+        print("Something wrong has occured")
